refact: move core to apps
