+++ bancodehoras/apps/core/static/img/icon-user.png
