+++ bancodehoras/apps/core/static/js/script.js
