+++ bancodehoras/apps/core/templates/modais/modal-mostrar-log.html
