+++ bancodehoras/bancodehoras/settings.py
@@ -19,8 +19,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'apps.core',
     'usuario',
-    'core',
     'movimentacao',
     'relatorio',
 ]
