+++ bancodehoras/bancodehoras/urls.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.contrib.auth.decorators import login_required
 from django.urls import path, include
 
-from core import views as views_core
+from apps.core.views import views as views_core
 from movimentacao import views as views_mov
 from relatorio import views as views_rel
 from usuario import views as views_usuario
